+++ WGU_Student_Mobile_App/ViewModels/InstructorsViewModel.cs
@@ -117,6 +117,12 @@ namespace WGU_Student_Mobile_App.ViewModels
             if (instructor == null)
                 return;
 
+            var result = await Application.Current.MainPage.DisplayAlert("Confirm Delete", "Are you sure you wish to delete this item", "OK", "Cancel");
+            if (!result)
+            {
+                return;
+            }
+
             instructorService.DeleteInstructor(instructor.Id);
             await Refresh();
         }
